+++ app/src/main/java/net/henryco/opalette/api/glES/render/graphics/shaders/textures/OPallTexture.java
@@ -284,8 +284,8 @@ public interface OPallTexture extends OPallShader {
 	OPallTexture setBitmap(Bitmap image, Filter filter);
 	OPallTexture setBitmap(Bitmap image);
 	OPallTexture setFlip(boolean x, boolean y);
-	OPallTexture setSize(int w, int h);
-	OPallTexture setRegion(int x, int y, int width, int height);
+	OPallTexture setSize(float w, float h);
+	OPallTexture setRegion(float x, float y, float width, float height);
 	OPallTexture setTextureDataHandle(int textureDataHandle);
 	OPallTexture setRotation(float angle);
 	int getTextureDataHandle();
